+++ R/trajPlot.R
@@ -131,7 +131,7 @@ trajPlot <- function(
   type = "default",
   group = NULL,
   cols = "default",
-  theme = "classic",
+  theme = "default",
   crs = 4326,
   map = TRUE,
   map.res = "medium",